requirements.txt
# Core dependencies
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# LangChain and AI
langchain==0.1.0
langchain-community==0.0.10
sentence-transformers==2.2.2
transformers==4.36.0
torch==2.1.0

# Vector Database
chromadb==0.4.22
faiss-cpu==1.7.4

# API and Web Framework
fastapi==0.108.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6

# UI
streamlit==1.29.0
streamlit-chat==0.1.1

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Utilities
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0
